test: move rule setup tests to separate file
